Make Process & Transition implement IEquatable<>
This is to get rid of comparer classes and to make the general
interface less noisy. From a theoretical point of view, saying
that processes must be equatable to other processes makes sense;
the same holds for transitions.

Internally, CcsProcess checks equality based on string
representations. Notice that the BuildRepr() methods are built
to “sort” subprocesses whenever needed, e.g. “a.0 + b.0” and
“b.0 + a.0” both get the same representation, namely “a.0+b.0”.